backup_current/requirements.txt
flask==3.0.3
flask-socketio==5.3.6
python-dotenv==1.0.1
google-api-python-client==2.149.0
google-auth==2.33.0
google-auth-httplib2==0.2.0
pandas==2.2.2
eventlet==0.36.1
requests==2.32.3
